writing each class in a separate file using webpack
